+++ src/Firebase/Exception/PermissionDenied.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Firebase\Exception;
+
+class PermissionDenied extends ApiException
+{
+
+}